+++ api-tests-karate/src/test/resources/logback-test.xml
@@ -8,7 +8,7 @@
     </appender>
   
     <appender name="FILE" class="ch.qos.logback.core.FileAppender">
-        <file>target/com.amido.stacks.api.karate.log</file>
+        <file>target/com.amido.stacks.tests.api.karate.log</file>
         <encoder>
             <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
         </encoder>
