--- api-tests/src/test/java/com/amido/stacks/api/OAuthConfigurations.java
@@ -1,19 +1,0 @@
-package com.amido.stacks.api;
-
-public enum OAuthConfigurations {
-  CLIENT_ID(System.getenv("CLIENT_ID")),
-  CLIENT_SECRET(System.getenv("CLIENT_SECRET")),
-  AUDIENCE(System.getenv("AUDIENCE")),
-  GRANT_TYPE(System.getenv("GRANT_TYPE")),
-  OAUTH_TOKEN_URL(System.getenv("OAUTH_TOKEN_URL"));
-
-  private final String config;
-
-  OAuthConfigurations(String config) {
-    this.config = config;
-  }
-
-  public String getOauthConfiguration() {
-    return config;
-  }
-}
