--- api-tests/src/test/java/com/amido/stacks/api/status/AppStatus.java
@@ -1,6 +1,0 @@
-package com.amido.stacks.api.status;
-
-public enum AppStatus {
-  RUNNING,
-  DOWN
-}
