+++ api-tests/src/test/java/com/amido/stacks/tests/api/OAuthConfigurations.java
@@ -1,0 +1,19 @@
+package com.amido.stacks.tests.api;
+
+public enum OAuthConfigurations {
+  CLIENT_ID(System.getenv("CLIENT_ID")),
+  CLIENT_SECRET(System.getenv("CLIENT_SECRET")),
+  AUDIENCE(System.getenv("AUDIENCE")),
+  GRANT_TYPE(System.getenv("GRANT_TYPE")),
+  OAUTH_TOKEN_URL(System.getenv("OAUTH_TOKEN_URL"));
+
+  private final String config;
+
+  OAuthConfigurations(String config) {
+    this.config = config;
+  }
+
+  public String getOauthConfiguration() {
+    return config;
+  }
+}
