+++ api-tests/src/test/java/com/amido/stacks/tests/api/status/AppStatus.java
@@ -1,0 +1,6 @@
+package com.amido.stacks.tests.api.status;
+
+public enum AppStatus {
+  RUNNING,
+  DOWN
+}
